refactor: utils -> entity.request 패키지 변경 -  `xxParser`로 끝나는 객체들이 request 하위의 기능이라는 판단
