+++ src/main/java/webserver/entity/request/RequestBody.java
@@ -3,8 +3,6 @@ package webserver.entity.request;
 import java.util.Collections;
 import java.util.Map;
 
-import webserver.utils.FormDataParser;
-
 public class RequestBody {
     private final Map<String, String> requestBody;
