+++ src/main/java/webserver/entity/request/RequestParser.java
@@ -1,4 +1,4 @@
-package webserver.utils;
+package webserver.entity.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
